+++ Extra/Templates/UIViewController + UITableView.xctemplate/___FILEBASENAME___ViewController.swift
@@ -3,7 +3,7 @@
 import UIKit
 
 // MARK: - ___VARIABLE_productName___ View Controller
-final class ___VARIABLE_productName___ViewController: UIViewController, ___VARIABLE_productName___Viewable, ___VARIABLE_productName___Navigatable, UITableViewDelegate, UITableViewDataSource {
+final class ___VARIABLE_productName___ViewController: UIViewController, ___VARIABLE_productName___Viewable, ___VARIABLE_productName___Navigable, UITableViewDelegate, UITableViewDataSource {
     // MARK: Subviews
     lazy var activityIndicator: UIActivityIndicatorView = initActivityIndicator()
 
@@ -46,7 +46,7 @@ final class ___VARIABLE_productName___ViewController: UIViewController, ___VARIA
 
     // MARK: Viewable
 
-    // MARK: Navigatable
+    // MARK: Navigable
 
     // MARK: Table View Delegate
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {