+++ Extra/Templates/UIViewController.xctemplate/___FILEBASENAME___ViewController.swift
@@ -3,7 +3,7 @@
 import UIKit
 
 // MARK: - ___VARIABLE_productName___ View Controller
-final class ___VARIABLE_productName___ViewController: UIViewController, ___VARIABLE_productName___Viewable, ___VARIABLE_productName___Navigatable {
+final class ___VARIABLE_productName___ViewController: UIViewController, ___VARIABLE_productName___Viewable, ___VARIABLE_productName___Navigable {
     // MARK: Subviews
     lazy var activityIndicator: UIActivityIndicatorView = initActivityIndicator()
 
@@ -46,6 +46,6 @@ final class ___VARIABLE_productName___ViewController: UIViewController, ___VARIA
 
     // MARK: Viewable
 
-    // MARK: Navigatable
+    // MARK: Navigable
 
 }